Introduces build metadata system All subprojects under the Revolution umbrella will now contain a build metadata file that is automatically generated during the compilation process.
This change offers two noteworthy benefits:
- It is possible to use factors like the branch and whether or not the changes have been committed to determine if
the build is intended for a development environment. In the future, this may or may not be expanded upon.
- Timestamps of the latest commit along with when the build was compiled to estimate how far into the development
process a particular build was
- Server owners/testers can use the /revolution command to obtain useful information about their build of a particular
subproject if they run into issues and need to report them

This will help identify what stage of development a build of a particular
plugin was from, how old a specific build is,